Revamp nested resource structure
I realized that my routes were becoming a little too nested for my
liking, so I consulted the documentation and made them shallow. Link:
https://guides.rubyonrails.org/routing.html#shallow-nesting.

This in itself causes a lot of tests and constants to be updated.
While doing this, some other minor changes were made:
- Added two test cases to the participants request spec, to cover
creation and deletion of a participant record. This was done because it
was realized that the participant controller's permitted params were
actually wrong and not permitting extras to go through.
- Fixed the participants controller participant params. They now allow
extras instead of the old participant_attributes.
- Fixed the concerns for finding resources. Namely, the `current_hunt`
method.
- Updated all the API modules and API calls in the frontend, after the
constants for the URLs were updated.