Revamp component structure, add roster details
The component structure and the fact that the old RosterParticipantsView
component was a mess indicated the need for a better structure.
This has been achieved by creating a new component with the
RosterDetails component, which is a parent of the Hunts component.

Additionally, I didn't like how the redirect was what provided the
roster, so it is much better now.