Add full licenses view, other supporting changes
Made the changes required in order to get proper license viewing with
pagination working nicely.

Some of the included changes:
- The ParticipantsTable now accepts any sort of object that has the
ParticipantBase properties.
- The ParticipantHandler takes several data transformation function, so
that many types of participant-like data can be swapped in instead of a
mere participant. Component which transform this data are called
adapters. There is an adapter for licenses, and an adapter for regular
participants.