Redesign workflow of components
The components in the app pages now have a common parent of the
UserDashboard, which is under the AppPage component itself.
I found this to be a lot cleaner, as the UserDashboard keeps track of
the current roster and current hunt, and more of the URLs are in one
place.

This also had the effect of fixing #17, as the duplicate request no
longer occurs.