Cleanup and modify action-cable client side
The client side of action cable is now a little cleaner, though still
with a few debug prints. I've switched it so that now it only uses the
official rails 6 client-side library (action-cable-react-jwt appeared to
be for rails 5).

Though it may now be a little less secure, so I'm looking into alternate
ways of handling authentication between the client and the server.