Fix CSRF token problems
As mentioned in #5, we were experiencing a lot of failures to verify the
CSRF tokens being passed with requests.

After some experimentation, we found out why, and now CSRF tokens are
passed to the frontend both as an embedded value in the page's HTML and
as a cookie.

The cookie is what the React frontend actually uses, since it will
update despite the cached pages.

Summary of changes:
- Updated frontend auth component to have promises with more meaningful
return values and notifications.
- ApiRequest module in frontend has new methods for getting the CSRF
token, since it is now in a cookie.
- Login component in frontend uses new Auth module hooks.
- Chance CSRF forgery strategy to exception, since this is stronger.
- ApplicationController now has a hook to update the CSRF cookie.

These changes fixed #5.