Add listener validation to AuthForm
It is now possible in the AuthForm to add a listener to another property
that will call the validation method on the subscriber. This is useful
for something like password confirmation, where the password
confirmation field's validator should be called either when the password
field changes or when the password confirmation field changes.