Implement Roster serializer for permissions
I've implemented a mechanism of serializing rosters and their child
models which always includes the permission, via the PermissionJson
concern.

The thing I don't like about it is that it will query for the permission
on each `as_json` call, which could be a big performance hit.

Hoping that the question I asked will be answered:
https://stackoverflow.com/questions/62419892/render-only-certain-associations-as-json-in-rails.