Refactor and improve roster loading
- Loading a roster is now completely handled by a separate component, to
make things cleaner.
- There is also a toplevel RosterContext component, responsible for
calling the LoadRoster component if the roster has not been loaded, or
calling the RosterDashboard when it is ready.
- Extracted the navbar on the top of the screen to a separate component,
the `UserNavbar`.
- Added functionality for child components of the UserDashboard to add
entries to the navbar dynamically. There is a problem with this
currently though: it doesn't have the roster when it's on other pages,
so the current roster and permissions tabs don't get populated.