Improve requests mechanisms
The way that error are now handled for requests is a little bit more
convenient. If the error is of the proper format, it is kept, else it is
made undefined for the actual frontend view parts and it is raised in
the console.

A lot of changes involved in doing this.