Add PaginatedLoader, change GenericPaginated
The PaginatedLoader component was created in order to have control over
teh state of the loaded items yet the convenience of having them loaded
for you.

The new GenericPaginated now encapsulates the PaginatedLoader and is the
same thing but with a default table and the state is no longer in
control of the caller component.

This all leverages the BlockLoader that was just created as well.