Add refresh permissions on change
The permissions adapter now updates the permission that is changed when
a request is made to change it.

There is a curreny known issue however: it will not update the current
owner's permission if the owner gets changed, in which case it would
look as though there are two owners at once until the user refreshses.