Add create, delete permissions
There is now a whole separate PermissionsPage component for modifying,
creating and deleting permissions.

For this reason, some of the previous functionality (like the select
form for a permission level) was extracted into separate components.

Permissions are now in a good enough state to continue with other stuff
if we want.