Refactor HuntDetails to use BlockLoader
Another component converted to BlockLoader.

I will say, I'm not sure that the fact that I had to type cast with
typescript is good. As the developer, I know that children of the
BlockLoader will have the state defined for them, but it would be nice
for typescript to know this too.

I could pass the state through the BlockLoader to the children as a
prop, but that feels wrong because if this were plain Javascript,
there's no way I'd do that. Don't want to change my runtime code for the
sake of compile time typescript checking.