Use container outside HuntNav, remove matchmake
Since matchmake is going to be done with a dialog instead, it will no
longer be a part of the HuntNav, nor will it be a direct route in the
HuntNavigator.

On top of this, I wanted the container rendering to be done outside of
the HuntNav component, so it is now done in the component that uses it,
the HuntNavigator.