Move filter state to URL for participants
The state of the currently selected hunt on the participant view is now
stored in the URL instead of as a variable.

This allows for linking to a specific hunt to filter by, which is
convenient for both the user and for development, since we can now
create links to a participant table that's been filtered.