Revamp routed modals, add add participant button
The routed modals and Hunt actions in general received a much needed
revamping, which makes them a lot more flexible thanks to the
modalPairGenerator factory that I created.

These changes were started as a result of trying to add a button link to
create licenses for the hunt, which has been accomplished with the
changes.