Add url-join for url joining
I grew tired of trying to do url joining myself, find it ridiculous that
javascript doesn't have this functionality built in, so I've added a
library in order to get the job done right.

I would make it myself, but there are a lot of edge cases and rules to
consider, much safer to use the library.