Disable create match button when empty arguments
Since it's no good for the user to be submitting invalid matchings, such
as when any of the matchings are empty, or when either of the pairs in a
matching are empty, we now check for such events and disable the
submission button if this is the case.

Additionally, a row is now added for the user on component mount, so
they do not have to add one immediately.