Make RoutedModal not assume partial path
As it was, the RoutedModal component assumed that the path being
provided was the end of the path. That assumption is no longer made at
the RoutedModal level, but is done instead in the HuntActions component.

This is all sort of pointless because we need to overhaul RoutedModal
with the new concept discovered when making MatchRoutedModal.