Hook up match view
The match view components have been hooked up to the main components so
that match viewing is now a thing.

Each match show view has an associated URL with it, that users can
navigate to straight away.

Additionally, when the user clicks the "View" button on a match in the
match list, it shows the match using the already loaded data from the
table. Subsequent navigation within the show match view however will
always load the match fresh, since there is no easy way of telling
whether or not it has been loaded, and even if it has been loaded the
data in the table could be out of date at that point.

Speaking of the data in the match list table being out of date, the
match list is loaded again when the modal is dismissed.

We may still want to improve the MatchDetails view, it's somewhat
barebones and possibly not informative enough, especially for something
that's supposed to show a single match in detail.