Fix search bar searching on previous value
There was a bug where the search bar was searching on the previously
inputted value (by one character) instead of the current value.

Reason for this bug had to do with state update, since the state
wouldn't have fully updated by the time the reference to the `onSearch`
function is made, so the solutions would be to either set up the timeout
in a useEffect hook that was subscribed to changes in the searchValue,
or to have the onSearch method get the value from the searchBar itself.
I chose the latter.