Integrate create-react-app
I've copied over a lot of files from the template generated with `yarn
create react-app --template typescript`, and the app runs with
typescript compilation errors.

What's left is fixing these errors, adding back eslint and prettier
configurations and then we should be in a good spot to begin work on the
backend.