Fix regression match view
After the changes in d037a63fc293cd776692aab58130e6f77c21e3e4, I added
`onReload` to the state watch in MatchRouteLoader.tsx. However, this
prop was being passed as an anonymous function, which would change
identity on each render.

I fixed this with usage of `useCallback` to keep the function's identity
the same.