Freshen up eslint config
After learning a bit from another project, I've decided to fix some
problems with the eslint config that I keep for this project.

It should now properly use the eslint configuration, as well as the
recommendations from the react plugins.

Additionally, the development server will not issue a compilation error
when eslint rules error out.