Add runtime environment variable loading
In order to have this application run as a Docker container in multiple
environments without having to rebuild the image, we needed a way to
configure the environment variables (in this case, just the URL for the
Rails server) at runtime, instead of at build time.

This has been achieved with a custom script which is invoked by the
nginx entrypoint when
the Docker container is created, and which generates a JavaScript file
which sets the variables to their proper values upon loading.

I think this can still be improved in a few ways however, notably by
handling cases where variables are unset in the generated JavaScript
module.